+++ src/test/java/faang/school/projectservice/mapper/StageInvitationMapperTest.java
@@ -1,7 +1,7 @@
 package faang.school.projectservice.mapper;
 
 import faang.school.projectservice.dto.stage_invitation.StageInvitationDto;
-import faang.school.projectservice.mapper.stage_invitation.StageInvitationMapperImpl;
+import faang.school.projectservice.mapper.StageInvitationMapperImpl;
 import faang.school.projectservice.model.TeamMember;
 import faang.school.projectservice.model.stage.Stage;
 import faang.school.projectservice.model.stage_invitation.StageInvitation;
